Move server side logging to common
The logging mechanism for the server is not really code specific to the
server.  We want to use it for the agent, so we are just doing a code
move first.
