+++ lib/pbench/server/report.py
@@ -10,8 +10,8 @@ import hashlib
 import socket
 from configparser import Error as NoSectionError, NoOptionError
 
+from pbench.common.logger import get_pbench_logger
 from pbench.server.indexer import PbenchTemplates, get_es, es_index, _op_type
-from pbench.server.logger import get_pbench_logger
 from pbench.server.utils import tstos
