+++ server/bin/pbench-backup-tarballs.py
@@ -9,7 +9,7 @@ import tempfile
 
 from pbench import PbenchConfig
 from pbench.common.exceptions import BadConfig
-from pbench.server.logger import get_pbench_logger
+from pbench.common.logger import get_pbench_logger
 from pbench.server.report import Report
 from pbench.server.s3backup import S3Config, Status, NoSuchKey
 from pbench.server.utils import md5sum, rename_tb_link, quarantine
