+++ server/bin/test-bin/test_logger_level.py
@@ -7,7 +7,7 @@ import sys
 
 from pbench import PbenchConfig
 from pbench.common.exceptions import BadConfig
-from pbench.server.logger import get_pbench_logger
+from pbench.common.logger import get_pbench_logger
 
 _NAME_ = "pbench-logger-level-test"
 cfg_name = os.environ["_PBENCH_SERVER_CONFIG"]
